jellyfish-shiproom/requirements.txt
# requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
jira>=3.5.2
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
google-auth>=2.23.4
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0